model, animationClip, BoneAnimation 경고를 지웠습니다.Model 에 머터리얼을 가변적으로 사용할수 있는 기능을 추가하였습니다.

+++ Engine_SOURCE/BoneAnimator.cpp
@@ -65,8 +65,8 @@ void BoneAnimator::LoadAnimations(const std::wstring& path)
 			if (sfileName.find("Demo") != std::string::npos)
 				continue;
 
-			startPos = sfileName.find_last_of("\\");
-			EndPos = sfileName.find_last_of(".");
+			startPos = static_cast<int>(sfileName.find_last_of("\\"));
+			EndPos = static_cast<int>(sfileName.find_last_of("."));
 
 			startPos += 1;
 			std::wstring wfileName = ConvertToW_String(sfileName.c_str());
